move function to helper file
